clean Family.jsx code a little bit + add family's transports feature
